backend/project-1/requirements.txt
sentence_transformers
langchain-huggingface
langchain_pinecone
python-dotenv
langchain
yfinance
requests
openai
flask
groq
bs4